### 2024-04-25 #6682 @editorialbot
[REVIEW]: RMAVIS v1.0: a Shiny application for the analysis of vegetation survey data and assignment to GB NVC communities
**Submitting author:** @ZekeMarshall (<a href="http://orcid.org/0000-0001-9260-7827">Zeke Marshall</a>)
**Repository:** https://github.com/NERC-CEH/RMAVIS
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0
