### 2024-04-25 #6682 @mikemahoney218
👋🏼 @ZekeMarshall, @adithirgis, @rasanderson this is the review thread for the paper. Just about all of our communications will happen here from now on. :smile:

```@editorialbot generate my checklist```

We aim for reviews to be completed within about 4 weeks. Please let me know if you require some more time.
