### 2024-04-25 #6682 @mikemahoney218
@ZekeMarshall , I just noticed that you changed the license of the project a few days ago to a non open-source license. JOSS only reviews projects that are open source using [an OSI-approved license](https://opensource.org/licenses), which none of the creative commons licenses are. Is there a reason you changed away from the MIT license? We will not be able to accept the project using the current license.
