### 2024-04-25 #6682 @ZekeMarshall
Hi @mikemahoney218 , apologies for this! I neglected to check that CCBY4.0 wasn't OSI-approved. I'll change this again in the morning when i'm in the office, likely to AAL, we just wanted to ensure attribution. PS thank you @adithirgis and @rasanderson for agreeing to review!
