### 2024-04-26 29f6ffe
Changing licence to LGPL-3.0, version 0.9995
