+++ NEWS.md
@@ -3,6 +3,12 @@ title: "Release notes for `RMAVIS`"
 output: html_document
 ---
 
+# RMAVIS 0.9995 (2024-04-26)
+*   Modifications:
+      -   Changing licence to the OSI-approved LGPL-3.0 licence to ensure compliance with JOSS.
+*   Fixes: NA
+*   New Features: NA
+
 # RMAVIS 0.9994 (2024-04-21)
 *   Modifications: NA
 *   Fixes: 
