+++ docker-compose.yml
@@ -4,7 +4,7 @@ services:
     build:
       context: .
       dockerfile: Dockerfile
-    image: "<container registry url>/rmavis:v0.9994"
+    image: "<container registry url>/rmavis:v0.9995"
     container_name: "rmavis"
     restart: always
     privileged: true
