+++ CITATION.cff
@@ -34,4 +34,4 @@ keywords:
   - NVC
   - Vegetation Science
   - Vegetation Survey
-license: CC BY 4.0
+license: LGPL-3.0
