### 2024-04-26 #6682 @ZekeMarshall
Hi @mikemahoney218 , I've now changed the licence to the OSI-approved LGPL 3.0 and made a new release. Could you set the version to v0.9995? Thanks!
