+++ JOSS/paper.bib
@@ -1,3 +1,20 @@
+@misc{allaire2023,
+  title      = {Rmarkdown: {{Dynamic Documents}} for {{R}}},
+  shorttitle = {Rmarkdown},
+  author     = {Allaire, {\relax JJ} and Xie, Yihui and Dervieux, Christophe and McPherson, Jonathan and Luraschi, Javier and Ushey, Kevin and Atkins, Aron and Wickham, Hadley and Cheng, Joe and Chang, Winston and Iannone, Richard},
+  year       = {2023},
+  abstract   = {Dynamic Documents for R},
+  copyright  = {GPL-3.0},
+  url        = {https://github.com/rstudio/rmarkdown}
+}
+
+@misc{attali2023,
+  title  = {Shinyjs: {{Easily Improve}} the {{User Experience}} of {{Your Shiny Apps}} in {{Seconds}}},
+  author = {Attali, Dean},
+  year   = {2023},
+  url    = {https://deanattali.com/shinyjs/}
+}
+
 @misc{bsbi2017ellenbergF,
   author = {BSBI},
   title  = {Ellenberg indicator F (moisture)},
@@ -147,6 +164,13 @@
   urldate    = {2024-02-20}
 }
 
+@misc{lin2024,
+  title  = {Reactable: {{Interactive Data Tables}} for {{R}}},
+  author = {Lin, Greg},
+  year   = {2024},
+  url    = {https://glin.github.io/reactable/}
+}
+
 @article{maciejewski2020,
   title      = {Vegetation Unit Assignments: Phytosociology Experts and Classification Programs Show Similar Performance but Low Convergence},
   shorttitle = {Vegetation Unit Assignments},
@@ -176,6 +200,85 @@
   publisher = {{Centre for Ecology \& Hydrology}},
   address   = {{Wallingford}},
   url       = {https://www.ceh.ac.uk/sites/default/files/Tablefit-Aug2016.zip}
+}
+
+@misc{meyer2024,
+  title  = {Shinybusy: {{Busy Indicators}} and {{Notifications}} for '{{Shiny}}' {{Applications}}},
+  author = {Meyer, Fanny and Perrier, Victor},
+  year   = {2024},
+  url    = {https://github.com/dreamRs/shinybusy}
+}
+
+@misc{nbn2024Anthocerotophyta,
+  author  = {{NBN Trust}},
+  title   = {Anthocerotophyta child taxa},
+  year    = {2024},
+  url     = {https://species.nbnatlas.org/species/NHMSYS0021386659#classification},
+  urldate = {2024-01-29}
+}
+
+@misc{nbn2024Bryophyta,
+  author  = {{NBN Trust}},
+  title   = {Bryophyta child taxa},
+  year    = {2024},
+  url     = {https://species.nbnatlas.org/species/NBNSYS0000172125#classification},
+  urldate = {2024-01-29}
+}
+
+@misc{nbn2024Marchantiophyta,
+  author  = {{NBN Trust}},
+  title   = {Marchantiophyta child taxa},
+  year    = {2024},
+  url     = {https://species.nbnatlas.org/species/NHMSYS0020796639#classification},
+  urldate = {2024-01-29}
+}
+
+@misc{nbn2024Pteridophyta,
+  author  = {{NBN Trust}},
+  title   = {Pteridophyta child taxa},
+  year    = {2024},
+  url     = {https://species.nbnatlas.org/species/NHMSYS0001747256#classification},
+  urldate = {2024-04-28}
+}
+
+@misc{nbn2024Tracheophyta,
+  author  = {{NBN Trust}},
+  title   = {Tracheophyta child taxa},
+  year    = {2024},
[…]
