+++ app.R
@@ -28,14 +28,13 @@ library(bookdown)
 library(tidyverse)
 library(plotly)
 library(kableExtra)
-library(janitor)
 library(writexl)
 
 # Ecology
 library(vegan)
 
 # Render documentation ----------------------------------------------------
-# rmarkdown::render(input = "docs/documentation.Rmd",  output_dir = "www")
+rmarkdown::render(input = "docs/documentation.Rmd",  output_dir = "www")
 
 # Source sub-modules ------------------------------------------------------
 source("modules/privacy_ui.R", local = TRUE)

[…]
