+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: RMAVIS
 Title: Assign Vegetation Survey Data to GB NVC Communities
-Version: 0.9995
+Version: 0.9996
 Date: 2024-04-30
 Authors@R: c(person(given = "Zeke",
                     family = "Marshall",
