+++ README.md
@@ -4,7 +4,7 @@
 <!-- badges: start -->
 
 [![Generic
-badge](https://img.shields.io/badge/Version-0.9995-green.svg)]()
+badge](https://img.shields.io/badge/Version-0.9997-green.svg)]()
 [![Project Status: Active – The project has reached a stable, usable
 state and is being actively
 developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
@@ -21,11 +21,5 @@ units, with additional exploratory analyses.
 ## Running the app
 
 You can run RMAVIS from [GitHub](https://github.com/ZekeMarshall/RMAVIS)
-by cloning the repository, calling `renv::restore()`, and then calling
-`shiny::runApp("app.R")`.
-
-If `renv::restore()` fails run
-`install.packages(unique(renv::dependencies()$Package), dependencies = TRUE)`,
-note that whilst the correct dependencies will be installed, the
-versions may not match those in the renv.lock file, and so the correct
-functioning of the app cannot be guaranteed.
+by cloning the repository, then calling `renv::activate()`,
+`renv::restore()`, and `shiny::runApp("app.R")`.
