### 2024-05-07 #6682 @ZekeMarshall
Hi @rasanderson & @adithirgis, I've bumped the version to v0.9997. This version set R 4.4 as the minimum version in response to the reported 'non-vulnerability' in R wherein arbitrary code can be executed when de-serialising R data objects, see [here](https://aitap.github.io/2024/05/02/unserialize.html), which is not in fact fixed by R 4.4.0.
