### 2024-05-20 #6682 @rasanderson
Hi @mikemahoney218 and @ZekeMarshall
I've now had a chance to look through the code, paper, and Shiny app., and think this is a really nice piece of work. I had a PhD student a couple of years ago doing lots of work on pseudoquadrats and MAVIS classifications (see [Butler, 2020 Chapter 3](https://libsearch.ncl.ac.uk/permalink/f/1jraif3/NCL_ALMA51188829440002411) ) so this is very interesting to read. There has been a need for this sort of app and functionality for some time. Some general points:

## The Shiny app
This worked well, both the online site hosted (presumably) by CEH, and the version built from the R code from the GitHub site. A number of relatively minor points:
- I did wonder whether the 'landing page' was right. When the app starts, the user is greeted by the data input page. It actually took me a couple of minutes before I noticed the 'Documentation' tab at the top right. I'd recommend a separate landing page to the data input page, and that the land page is user-friendly, so that it is obvious that the data needs to be input / loaded, then options chosen etc.
- Maybe change 'Documentation' to 'Help'.
- On both my PCs, a centrally-managed Windows PC, and a Linux workstation, the Documentation comes up in a relatively small window rather than the full height of the screen. Not quite sure why this is, but possibly a 'FluidRow' shiny control needs modification in one of the R files, though I haven't managed to track it down.
- Similarly, some of the 'popup' boxes have a close button, but this is a long way down, and you have to scroll down to find it. e.g. after clicking 'Validate survey data' a box headed 'Validate Survey Table Data' is displayed, and the background of the Shiny app becomes blurred. The Close button is not displayed, as it is off the bottom of the screen: I seemed to be doing a bit of random clicking before I found what was going on, and realised that I had to scroll down the full length of the screen.
- How easy would it be to add other data upload format options? e.g. TurboVeg format, although I realise that as a commercial package this might not be compatible.
- I ran some of the example datasets, generated outputs, downloaded reports, and these were all great.

## Paper
This provides a good summary of the needs for the work, is clearly written and correctly referenced. Main queries are:
- The paper (and Shiny app) is very UK-focussed, which to some extent is inevitable given its dependence on the National Vegetation Classification. JOSS is an international journal, however, so I think it's important to acknowledge a wider applicability of this type of method in countries other than the UK and Ireland. We had a similar NVC-based paper published a couple of years ago [Butler & Sanderson, 2022](https://doi.org/10.1111/1365-2664.14166) and reported examples from USA, Canada, Australia, NZ, European Vegetation Archive etc. to try and ensure that an international readership would still find the research relevant.
- Both your paper and the Shiny app make use of pseudoquadrats, which I originally developed in collaboration with Steve Rushton and Nigel Critchley in the 1990s [Sanderson et al 1995](https://doi.org/10.1016/S0301-4797(95)90426-3) [Critchley et al 2002](https://doi.org/10.1016/S0006-3207(01)00183-5). However, the technique has never been widely adopted, and I suspect will be very unfamiliar to many readers, especially in an international audience. I'd therefore recommend that you expand your explanation of the pseudoquadrat concept, on both the Shiny app and in the paper, to ensure no ambiguity.
- Use of Decorana (DCA). This is quite a controversial area: both TWINSPAN and DCA were the default multivariate classification and ordination systems respectively when the NVC was developed, but both have fallen out of favour since then. DCA has been criticised for sometimes over-correcting the 'arch' effect but then introducing a 'wedge' or 'triangle' effect in the display of points. You've used Oksanen's version of DCA, which will ensure that the results aren't sensitive to the input order of quadrats, which was a problem in the original version. However, in Numerical Ecology [Legendre & Legendre 2012](https://shop.elsevier.com/books/numerical-ecology/legendre/978-0-444-53868-0) they caution that '...MDS is the best technique for recovering complex gradients, not DCA', and '...detrending should be avoided, except for the specific purpose of estimating the lengths of gradients', but that 'satisfactory mathematical solutions to the problem of detrending remain to be found'. Of course, given the close association of the NVC with these methods, I think you're right to present DCA. However, I wonder whether you could modify the Shiny interface so that alternative MVA techniques could be offered to the end-user, such as NMDS, or PCA and its constrained equivalent, RDA etc? These are all very simple to implement in the R vegan package.

## R code and Github site
Again, relatively minor comments.
- The GitHub site could be improved. I think you need a much more thorough README.md file than the one currently provided. It's likely that quite a number of readers and / or eventual users would like more information on the GitHub site. When I am installing new software from GitHub, I much prefer a comprehensive README.md (even it some of its contents are repeated later, in the app), than a rather sparse one.
- The R code is excellent, well-structured and organised, using Shiny modules to control the operation of the app, and lots of comments throughout.
- Use of renv is a good idea. My Linux box automatically installed renv and prepared the R project correctly. My (centrally managed) Windows machine has an old version of RStudio (2022.12.0), and it failed to detect renv when I opened the R project file. You may therefore want to add instructions about installing renv, or specifiy a minimum version of RStudio for use with the app.
- There isn't a formal API for this set of code, as it's an R project to drive a shiny app. However, to maximise takeup internationally, so that it can be used for other vegetation classifications not just the NVC, it needs to be easily customised. Having Shiny modules helps this. However, I notice all your R functions, for example in the six files in the R folder, are fully documented in the code. Once the project is loaded, it means that doing the usual `?habCor_data` brings up the help information for that function. I'm wondering whether you've thought of making this a formal R package, as well as being a shiny app? This would expose the call, parameters, return values of these functions more explicitly. I'm not sure how complex it would be to create it as an R package hosted on GitHub (it doesn't need to go onto CRAN), but as an R library you'd then also have the possibility of a reference manual, package vignettes etc. Again, I'm thinking in terms of ensuring that what to me is an excellent tool has as wide an uptake as possible outside of the UK, by allowing end-users to easily customise parts of what you've created for their own shiny apps for their own regional vegetation classifications.

Enjoyed reading the research you've been doing - great stuff.
Roy
