### 2024-05-20 NERC-CEH/RMAVIS#1 @ZekeMarshall
Add landing page
Add a home/landing page containing an overview of RMAVIS. In response to JOSS review comments from Roy Sanderson (https://github.com/openjournals/joss-reviews/issues/6682).
