### 2024-05-20 NERC-CEH/RMAVIS#2 @ZekeMarshall
Improve modal popups
Change the position of the modal pop-up close buttons as they are not visible on screen when the modal pop-ups are longer than the page. In response to JOSS review comments from Roy Sanderson (https://github.com/openjournals/joss-reviews/issues/6682).
