### 2024-05-20 NERC-CEH/RMAVIS#3 @ZekeMarshall
Improve documentation of pseudo-quadrat method
Improve the quality of the documentation regarding the generation and use of pseudo-quadrats. In response to JOSS review comments from Roy Sanderson (https://github.com/openjournals/joss-reviews/issues/6682).
