### 2024-05-20 NERC-CEH/RMAVIS#4 @ZekeMarshall
Instruct users to install {renv}
Instruct RMAVIS users to install {renv} in the README and app.R file. In response to JOSS review comments from Roy Sanderson (https://github.com/openjournals/joss-reviews/issues/6682).
