+++ JOSS/paper.bib
@@ -551,6 +551,15 @@
   pages     = {203-214}
 }
 
+@misc{xie2023,
+  title      = {Markdown: {{Render Markdown}} with 'Commonmark'},
+  shorttitle = {Markdown},
+  author     = {Xie, Yihui},
+  year       = {2023},
+  month      = dec,
+  url        = {https://cran.r-project.org/web/packages/markdown/index.html}
+}
+
 @misc{xie2024,
   title     = {Bookdown: {{Authoring Books}} and {{Technical Documents}} with {{R Markdown}}},
   author    = {Xie, Yihui},
