+++ modules/documentation_ui.R
@@ -9,9 +9,7 @@ documentationUI <- function(id) {
       shiny::div(
         htmltools::tags$iframe(src = "documentation.html",
                                width = '100%',
-                               height = "600px",
-                               style = "border:none;")
-        #, style = "margin-right: 5px !important;"
+                               style = "border:none; height: 85vh;")
       )
     )
   )
