### 2024-05-22 #6682 @ZekeMarshall
Hi @rasanderson ,

Thank you for your positive review!

In response to your comments:

**The Shiny App**

- Good point about a landing page, I've added a landing page and rendered the README to this page.
- Thanks for pointing out that the documentation opens in a small window, I've fixed the height to 85% of the screen, so it should scale for different monitors now.
- I've added the ability to close the modal pop ups by clicking outside the window. Thanks for this suggestion.
- This isn’t something we plan on adding at the moment, as we’re trying not to build out too much data wrangling functionality, rather leaving the user to do some work to get it into our simple 5 column format. But it is something that we could do in the future, especially given that packages such as {vegtable} already exist.

**Paper**

- Thanks for pointing me in the direction of Butler & Sanderson (2022)!  I did set up the app so that we could facilitate the incorporation of other vegetation classifications and taxonomic backbones in the future, as all underlying data are initialised as setup data in the `setupData_server.R` file. So I’m open to the idea, but for v1.0 I’d like to stick to the GB focus, perhaps building out this functionality for a future release.
- Thanks for pointing me in the direction of Sanderson et al (1995) and Critchley et al (2002), I agree we should provide a better description, I’ve added these references to the documentation!
- Good point RE the DCA. This is definitely something I thought/I’m currently thinking about, but not for v1.0 I think.

**R code and Github site**

- Thanks for the suggestion, I've improved the README!
- I’ll check out what the minimum RStudio version would be/is. However, I use Visual Studio Code and RStudio interchangeably so I also wouldn’t want to be too prescriptive about the choice of IDE. I have now instructed the user to install {renv} in the README.
- I have structured the repo as a basic R package, which can be installed from Github with `remotes::install_github(“NERC-CEH/RMAVIS”)`, and I use it in a few other projects now. So I will definitely be maintaining it as a package!

As I mentioned before, the initial set up to generalise RMAVIS has been done in “setupData_server.R”, but I should also document the modules and export them as functions to facilitate re-use by others.

This may not be for v1 though, as it would require more thought and building in the facility to turn off modules such as the MVA modules if no pre-calculated ordinations are available for the “National Reference” MVA; or when EIVs aren’t available etc, along with restructuring and renaming the bundled GB-specific data.

So the bundled data and functions in /R are available via the RMAVIS ‘package’, but the shiny code is not yet!

Thanks again, the changes I've made above are contained in the latest release (v0.9998)!

Best regards,

Zeke
