### 2024-06-03 #6682 @mikemahoney218
@adithirgis , I wanted to reach back out to ask how things are progressing, and to see if you had any questions/comments about your review for JOSS!

Just as a reminder, the first step in the review is to post `@editorialbot generate my checklist` as the start of a new comment in this thread, which will generate a review checklist for you to use.

Thanks!
