### 2024-06-15 #6682 @adithirgis
Hello @ZekeMarshall,

A great job on this software. A few minor points to share, please correct me if any points are already implemented -

### Documentation
- Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software - I think this needs to be separate document something like this - https://github.com/r-lib/usethis/blob/main/.github/CONTRIBUTING.md

### Functionality

- Installation: Does installation proceed as outlined in the documentation? - If I use the ```remotes::install_github("NERC-CEH/RMAVIS"); library(RMAVIS)``` what are the next steps to run the app. Please add it in the documentation. The other options works well.
- I see that the online version uploaded here - https://connect-apps.ceh.ac.uk/RMAVIS/ works similar to the local version that is very well and smooth (the local version takes time to setup though), but I was thinking most of the data seems to be from UK. Since this app can be used in other
- Also I recommend using devtools::check() - there are a few warnings and notes to consider.
- Download report shows this error.
```
Warning in system2(..., stdout = if (use_file_stdout()) f1 else FALSE, stderr = f2) :
  '"pdflatex"' not found
Warning: Error in : LaTeX failed to compile C:\Users\adith\AppData\Local\Temp\RtmpcNGUY2\file32643bc25352.tex. See https://yihui.org/tinytex/r/#debugging for debugging tips.
  1: shiny::runApp
```
