+++ JOSS/paper.bib
@@ -335,6 +335,15 @@
   doi        = {10.1046/j.1365-2664.2002.00718.x}
 }
 
+@article{raper2015,
+  title = {{{UK Species Inventory}} - {{Master}} Copy},
+  author = {Raper, Chris},
+  year = {2015},
+  month = jan,
+  publisher = {Natural History Museum},
+  doi = {10.5519/0038741}
+}
+
 @article{ridding2020,
   title     = {Long-term change in calcareous grassland vegetation and drivers over three time periods between 1970 and 2016},
   volume    = {221},
