### 2024-06-16 d21ef32
Streamlining dependencies, preparing for v1
