+++ JOSS/paper.bib
@@ -98,6 +98,15 @@
   url    = {https://shiny.posit.co/}
 }
 
+@misc{cheng2024,
+  title      = {Htmltools: {{Tools}} for {{HTML}}},
+  shorttitle = {Htmltools},
+  author     = {Cheng, Joe and Sievert, Carson and Schloerke, Barret and Chang, Winston and Xie, Yihui and Allen, Jeff and Software, Posit and PBC},
+  year       = {2024},
+  month      = apr,
+  copyright  = {GPL-2 {\textbar} GPL-3 [expanded from: GPL ({$\geq$} 2)]}
+}
+
 @misc{cieem2022,
   author       = {CIEEM},
   howpublished = {PDF},
@@ -336,12 +345,21 @@
 }
 
 @article{raper2015,
-  title = {{{UK Species Inventory}} - {{Master}} Copy},
-  author = {Raper, Chris},
-  year = {2015},
-  month = jan,
+  title     = {{{UK Species Inventory}} - {{Master}} Copy},
+  author    = {Raper, Chris},
+  year      = {2015},
+  month     = jan,
   publisher = {Natural History Museum},
-  doi = {10.5519/0038741}
+  doi       = {10.5519/0038741}
+}
+
+@manual{rcoreteam,
+  title        = {R: A Language and Environment for Statistical Computing},
+  author       = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address      = {Vienna, Austria},
+  year         = {2024},
+  url          = {https://www.R-project.org/}
 }
 
 @article{ridding2020,
@@ -448,6 +466,15 @@
   isbn      = {978-0-429-44727-3}
 }
 
+@misc{sievert2023,
+  title      = {Bsicons: {{Easily Work}} with '{{Bootstrap}}' {{Icons}}},
+  shorttitle = {Bsicons},
+  author     = {Sievert, Carson and Posit Software PBC and Otto, Mark},
+  year       = {2023},
+  month      = nov,
+  copyright  = {MIT + file LICENSE}
+}
+
 @misc{sievert2024,
   title  = {Bslib: {{Custom}} '{{Bootstrap}}' '{{Sass}}' {{Themes}} for 'Shiny' and 'Rmarkdown'},
   author = {Sievert, Carson and Cheng, Joe and {Aden-Buie}, Garrick},
@@ -491,6 +518,15 @@
   year      = {2024}
 }
 
+@misc{stefan2022,
+  title      = {Magrittr: {{A Forward-Pipe Operator}} for {{R}}},
+  shorttitle = {Magrittr},
+  author     = {Stefan, Stefan Milton and Wickham, Hadley and Henry, Lionel and RStudio},
+  year       = {2022},
+  month      = mar,
+  copyright  = {MIT + file LICENSE}
+}
+
 @article{sturbois2023,
   title      = {Ecological Quality Assessment: {{A}} Framework to Report Ecosystems Quality and Their Dynamics from Reference Conditions},
   shorttitle = {Ecological Quality Assessment},
@@ -502,6 +538,15 @@
   pages      = {e4726},
   issn       = {2150-8925},
   doi        = {10.1002/ecs2.4726}
+}
+
+@misc{vaidyanathan2023,
+  title      = {Htmlwidgets: {{HTML Widgets}} for {{R}}},
+  shorttitle = {Htmlwidgets},
+  author     = {Vaidyanathan, Ramnath and Xie, Yihui and Allaire, J. J. and Cheng, Joe and Sievert, Carson and Russell, Kenton and Hughes, Ellis and Software, Posit and PBC},
+  year       = {2023},
+  month      = dec,
+  copyright  = {MIT + file LICENSE}
 }
[…]
