### 2024-06-16 #6682 @ZekeMarshall
Hi @adithirgis ,

Thanks for your review!!

In response to your comments I have pushed a provisional v1.0 release to the v1 branch, this branch incorporates the following changes:

- Added a function `RMAVIS::runApp()` to run RMAVIS after installation as a package.
- Added instructions on how to run the `RMAVIS` app using the function above in the README.
- Added instructions on how to install a LaTeX distribution to ensure the report can be generated (specifically using `{tinytex}`) in the README.
- Included contributor guidelines.
- Streamlined the dependencies.

In terms of the package checks, ensuring there are no warnings is a tricky job. The biggest issue is the presence of non-ASCII strings in the taxonomic backbone data. I need to think about the best way to handle this as I need to maintain compatibility with the taxonomic backbone source and convert these non-ASCII strings to ASCII for `RMAVIS` and check they are all correct. I would also like to do this following a major update to the United Kingdom Species Inventory later this year. @mikemahoney218 are you happy for me to leave this job for a future release?

Lastly, @adithirgis the only comment I'm unclear about is in the **Functionality** section of your review - "Since this app can be used in other". What do you mean by this?

Thanks again! Please let me know if my changes have addressed your comments. When everyone is happy I will then release the actual v1.0.

Best regards,

Zeke
