+++ app.R
@@ -2,4 +2,3 @@
 dir <- system.file("app", package = "RMAVIS")
 setwd(dir)
 shiny::shinyAppDir(".")
-RMAVIS::runApp()
