+++ JOSS/paper.bib
@@ -218,6 +218,19 @@
   url    = {https://github.com/dreamRs/shinybusy}
 }
 
+@article{mucina2016a,
+  title = {Vegetation of {{Europe}}: Hierarchical Floristic Classification System of Vascular Plant, Bryophyte, Lichen, and Algal Communities},
+  shorttitle = {Vegetation of {{Europe}}},
+  author = {Mucina, Ladislav and B{\"u}ltmann, Helga and Dier{\ss}en, Klaus and Theurillat, Jean-Paul and Raus, Thomas and {\v C}arni, Andra{\v z} and {\v S}umberov{\'a}, Kate{\v r}ina and Willner, Wolfgang and Dengler, J{\"u}rgen and Garc{\'i}a, Rosario Gavil{\'a}n and Chytr{\'y}, Milan and H{\'a}jek, Michal and Di Pietro, Romeo and Iakushenko, Dmytro and Pallas, Jens and Dani{\"e}ls, Fred J.A. and Bergmeier, Erwin and Santos Guerra, Arnoldo and Ermakov, Nikolai and Valachovi{\v c}, Milan and Schamin{\'e}e, Joop H.J. and Lysenko, Tatiana and Didukh, Yakiv P. and Pignatti, Sandro and Rodwell, John S. and Capelo, Jorge and Weber, Heinrich E. and Solomeshch, Ayzik and Dimopoulos, Panayotis and Aguiar, Carlos and Hennekens, Stephan M. and Tich{\'y}, Lubom{\'i}r},
+  year = {2016},
+  journal = {Applied Vegetation Science},
+  volume = {19},
+  number = {S1},
+  pages = {3--264},
+  issn = {1654-109X},
+  doi = {10.1111/avsc.12257}
+}
+
 @misc{nbn2024Anthocerotophyta,
   author  = {{NBN Trust}},
   title   = {Anthocerotophyta child taxa},
