### 2024-07-01 #6682 @adithirgis
Hello,

Thank you for your response. I see that the changes are implemented.
Regarding - "Since this app can be used in other" I meant in other _countries_ as well.

Last question now that we need ```install.packages("tinytex")``` is it a required package in DESCRIPTION? Apologies I am confused regarding this step. Rest everything looks good! Good luck with the package.

Thanks,
Adithi
