### 2024-07-02 #6682 @ZekeMarshall
Hi @adithirgis ,

Thanks for your reply and checking the v1 branch!

I've added {tinytex} in the suggests section of the DESCRIPTION file, as the user may already have a LaTeX distribution installed and so may not need {tinytex}. Thanks for pointing this out!

@mikemahoney218 , if you are ok with the reviews from @adithirgis and @rasanderson am I now ok to take the following steps for potential acceptance?

1. Merge v1 into main and release v1.0.0
2. Re-generate the article pdf

Cheers,

Zeke
