+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: RMAVIS
 Title: Assign Vegetation Survey Data to GB NVC Communities
 Version: 1.0
-Date: 2024-06-16
+Date: 2024-07-02
 Authors@R: c(person(given = "Zeke",
                     family = "Marshall",
                     role = c("aut", "cre"),
@@ -56,7 +56,8 @@ Imports:
   vegan,
   writexl
 Suggests:
-  testthat
+  testthat,
+  tinytex
 URL: https://github.com/NERC-CEH/RMAVIS
 BugReports: https://github.com/NERC-CEH/RMAVIS/issues
 Depends: 
