### 2024-07-02 9e2ea3b
Re-generating paper.md
