+++ JOSS/paper.md
@@ -25,7 +25,7 @@ affiliations:
     index: 2
   - name: School of Environmental Sciences, University of Liverpool, Liverpool, L69 3GP, UK
     index: 3
-date: "2024-06-16"
+date: "2024-07-02"
 bibliography: paper.bib
 output:
   html_document:
