### 2024-07-02 #6682 @mikemahoney218
Thank you so much @rasanderson and @adithirgis for your thoughtful reviews here.

@ZekeMarshall, at this point could you follow the checklist in the above message (and yes, merge & release 1.0.0)?

As a general note -- I'm generally offline this week for a holiday, so might be slow to respond to this issue. I'll be back next week and only intermittently online until then.
