### 2024-07-02 #6682 @ZekeMarshall
Hi @mikemahoney218 ,

Thanks!

I can confirm that I have completed the tasks you mention above.

The version number is **v1.0**.

The archive DOI for the **v1.0** release is **10.5281/zenodo.12627165**

The archive DOI for all releases is **10.5281/zenodo.10818640**

Please let me know if there is anything else you require!

PS. I accidentally created duplicate releases on Zenodo for v1.0. I have requested deletion of the duplicates, this won't affect the DOI of the v1.0 release mentioned above (10.5281/zenodo.12627165) - apologies if this is an inconvienience!

Best regards,

Zeke
