### 2024-07-08 #6682 @mikemahoney218
@ZekeMarshall , a few paper comments:

* Line 11: this is _such_ a nitpick, but it's technically either "Shiny" or "Shiny for R"; R Shiny is popular but technically incorrect. I'd suggest changing to either "Shiny" or "Shiny for R"
* Line 29: Same comment, but I think it should be "_an_ R Shiny" if not changing
* CRAN packages all got DOIs two-ish weeks ago; would you please update your citations to include the new DOIs? Most packages have their DOI in the table above, except a few
* Related -- could you cite the CRAN version (with DOI) of kableExtra, Shiny, RMarkdown
* Can you check that you've got things in your Bibtex bracketed so that they're capitalized appropriately? Specific notes:
  *  Most packages should be all lowercase (so `title = {{rmarkdown}: Dynamic Documents for R}` and so on):
    * rmarkdown
    * shinyjs
    * htmltools
    * shinybusy
    * writexl
    * rhandsontable
    * magrittr
    * htmlwidgets
    * markdown
    * bookdown
    * Shiny itself is an exception; Posit capitalizes it throughout the documentation
  * I think the Ellenberg indicators should be capitalized?
  * The "R" in package names needs to be capitalized:
    * Shiny: Web application framework for r
  * Sievert 2020, 2024, 2023 are all capitalized differently than other entries -- might have too many `{}`
    * Same with Xie 2023, 2024
* Sivert 2023 -- is Posit listed as an author? On purpose? If so, needs `{}` to make sure it's listed as `{Posit, PBC}` -- right now it looks like bizarre initials
* If any of the `No DOI given, and none found` above actually _do_ have a DOI, it'd be awesome if we can add it -- but if not, no worries, not a blocker on our end

Thanks!
