### 2024-07-08 #6682 @ZekeMarshall
Hi @mikemahoney218 , thanks for reviewing the paper! I'll work through these comments this afternoon and get back to you asap. One question, I've bundled the JOSS paper in the main branch and released v1, so if I update the paper it will be slightly out-of-sync with the Zenodo release, I think the best thing to do is to create a new branch called MM_JOSS_edits and update the paper there, is this ok with you?
