### 2024-07-08 #6682 @mikemahoney218
Sorry, I should have mentioned this -- from our perspective, the archive is of _the software_ (as it was when we reviewed it), because we as the journal are responsible for archiving the paper. As such, changes to the paper (and other non-software bits) don't need any changes to the archive or anything like that -- we'll create the authoritative archive of the paper when we move to accept. So it should be fine to make changes on main.
