### 2024-07-08 #6682 @ZekeMarshall
Hi Mike, ok cool thanks! Do you have any recommendations for persons wanting to store the JOSS paper in the main branch and keep Github and Archive releases in sync with the published release version as stipulated in the paper. I can see how there may be a bit of a chicken and egg situation with the editor setting the version and archive, but then there been further changes prior to acceptance. Is it just best practice to store the paper in a separate branch all together? Apologies if i've missed something and thanks again!
