+++ JOSS/paper.bib
@@ -1,23 +1,37 @@
-@misc{allaire2023,
-  title      = {Rmarkdown: {{Dynamic Documents}} for {{R}}},
-  shorttitle = {Rmarkdown},
-  author     = {Allaire, {\relax JJ} and Xie, Yihui and Dervieux, Christophe and McPherson, Jonathan and Luraschi, Javier and Ushey, Kevin and Atkins, Aron and Wickham, Hadley and Cheng, Joe and Chang, Winston and Iannone, Richard},
-  year       = {2023},
-  abstract   = {Dynamic Documents for R},
-  copyright  = {GPL-3.0},
-  url        = {https://github.com/rstudio/rmarkdown}
+@misc{allaire2024,
+  title     = {{rmarkdown}: {{Dynamic Documents}} for {{R}}},
+  author    = {Allaire, J. J. and Xie, Yihui and Dervieux, Christophe and McPherson, Jonathan and Luraschi, Javier and Ushey, Kevin and Atkins, Aron and Wickham, Hadley and Cheng, Joe and Chang, Winston and Iannone, Richard and Dunning, Andrew and Yasumoto, Atsushi and Schloerke, Barret and Sievert, Carson and Ryan, Devon and Aust, Frederik and Allen, Jeff and Seo, JooYoung and Barrett, Malcolm and Hyndman, Rob and Lesur, Romain and Storey, Roy and Arslan, Ruben and Oller, Sergio and PBC, Posit and Otto, Mark and Thornton, Jacob and Farkas, Alexander and Jehl, Scott and Sagalaev, Ivan and Franko, Greg and MacFarlane, John and Raggett, Dave and Gandy, Dave and Sperry, Ben and Drifty and Lister, Aidan and Jonsson, Benct Philip and Krewinkel, Albert},
+  year      = {2024},
+  month     = may,
+  urldate   = {2024-07-08},
+  abstract  = {Convert R Markdown documents into a variety of formats.},
+  copyright = {GPL-3},
+  doi       = {10.32614/CRAN.package.rmarkdown}
 }
 
-@misc{attali2023,
-  title  = {Shinyjs: {{Easily Improve}} the {{User Experience}} of {{Your Shiny Apps}} in {{Seconds}}},
-  author = {Attali, Dean},
-  year   = {2023},
-  url    = {https://deanattali.com/shinyjs/}
+@misc{attali2021,
+  title     = {{shinyjs}: {{Easily Improve}} the {{User Experience}} of {{Your Shiny Apps}} in {{Seconds}}},
+  author    = {Attali, Dean},
+  year      = {2021},
+  month     = dec,
+  urldate   = {2024-07-08},
+  copyright = {MIT + file LICENSE},
+  doi       = {10.32614/CRAN.package.shinyjs}
+}
+
+@misc{bache2022a,
+  title     = {{magrittr}: {{A Forward-Pipe Operator}} for {{R}}},
+  author    = {Bache, Stefan Milton and Wickham, Hadley and Henry, Lionel and PBC, Posit},
+  year      = {2022},
+  month     = mar,
+  urldate   = {2024-07-08},
+  copyright = {MIT + file LICENSE},
+  doi       = {10.32614/CRAN.package.magrittr}
 }
 
 @misc{bsbi2017ellenbergF,
   author = {BSBI},
-  title  = {Ellenberg indicator F (moisture)},
+  title  = {Ellenberg indicator {F} (moisture)},
   day    = {15},
   month  = {10},
   year   = {2017},
@@ -26,7 +40,7 @@
 
 @misc{bsbi2017ellenbergL,
   author = {BSBI},
-  title  = {Ellenberg indicator L (light)},
[…]

[…]
