### 2024-07-08 #6682 @ZekeMarshall
> > Do you have any recommendations for persons wanting to store the JOSS paper in the main branch and keep Github and Archive releases in sync with the published release version as stipulated in the paper
>
> So from my perspective, the Zenodo archive version we've got should reflect the software at time of acceptance, and the JOSS archive should reflect the paper at time of acceptance. If someone is looking to cite the paper, they should be citing the JOSS DOI -- so if you _really_ wanted to make sure things didn't drift from each other, you could remove the paper from the Zenodo archive. You can also cut a new release, and get a new Zenodo archive, once we've finished and accepted everything -- and could even include the final PDF with JOSS DOI and so on in that release.
>
> Otherwise, though, I think this is something that will cause a lot of friction and not return a lot of benefit; the paper and JOSS materials aren't part of your software, so the Zenodo archive doesn't need to reflect the final reviewed version of those. JOSS will create a permanent archive of the paper and its authoritative DOI, and the Zenodo archive is simply a link from the website so people can see what the software looked like at time of acceptance.

Thanks for the explanation! I think this synchronisation issue is tricky given that JOSS explicitly recommends placing the paper in the same repository as the software. I might place the paper in a different repo next time.
