### 2024-07-08 #6682 @danielskatz
If you put the paper in a different repo than the software, the JOSS commands that look at the paper and the software won't both work.
