### 2024-07-08 #6682 @mikemahoney218
@ZekeMarshall I think the better option, if you don't want the paper in the archive, is to not rely on Zenodo's automatic GitHub archiving plugin for JOSS. You can create your own archival upload manually and include only the software files, and use that archive for JOSS. (You could also do this and still use the automatic plugin for other releases.) The automatic tool is great, but not necessarily the best fit for our process here!
