### 2024-07-08 #6682 @ZekeMarshall
I've just noticed that the "Posit PBC" author is not rendering correctly here (it was locally sorry), I'll try and fix this now and push again.
