+++ JOSS/paper.bib
@@ -1,6 +1,6 @@
 @misc{allaire2024,
   title     = {{rmarkdown}: {{Dynamic Documents}} for {{R}}},
-  author    = {Allaire, J. J. and Xie, Yihui and Dervieux, Christophe and McPherson, Jonathan and Luraschi, Javier and Ushey, Kevin and Atkins, Aron and Wickham, Hadley and Cheng, Joe and Chang, Winston and Iannone, Richard and Dunning, Andrew and Yasumoto, Atsushi and Schloerke, Barret and Sievert, Carson and Ryan, Devon and Aust, Frederik and Allen, Jeff and Seo, JooYoung and Barrett, Malcolm and Hyndman, Rob and Lesur, Romain and Storey, Roy and Arslan, Ruben and Oller, Sergio and PBC, Posit and Otto, Mark and Thornton, Jacob and Farkas, Alexander and Jehl, Scott and Sagalaev, Ivan and Franko, Greg and MacFarlane, John and Raggett, Dave and Gandy, Dave and Sperry, Ben and Drifty and Lister, Aidan and Jonsson, Benct Philip and Krewinkel, Albert},
+  author    = {Allaire, J. J. and Xie, Yihui and Dervieux, Christophe and McPherson, Jonathan and Luraschi, Javier and Ushey, Kevin and Atkins, Aron and Wickham, Hadley and Cheng, Joe and Chang, Winston and Iannone, Richard and Dunning, Andrew and Yasumoto, Atsushi and Schloerke, Barret and Sievert, Carson and Ryan, Devon and Aust, Frederik and Allen, Jeff and Seo, JooYoung and Barrett, Malcolm and Hyndman, Rob and Lesur, Romain and Storey, Roy and Arslan, Ruben and Oller, Sergio and {Posit PBC} and Otto, Mark and Thornton, Jacob and Farkas, Alexander and Jehl, Scott and Sagalaev, Ivan and Franko, Greg and MacFarlane, John and Raggett, Dave and Gandy, Dave and Sperry, Ben and Drifty and Lister, Aidan and Jonsson, Benct Philip and Krewinkel, Albert},
   year      = {2024},
   month     = may,
   urldate   = {2024-07-08},
@@ -21,7 +21,7 @@
 
 @misc{bache2022a,
   title     = {{magrittr}: {{A Forward-Pipe Operator}} for {{R}}},
-  author    = {Bache, Stefan Milton and Wickham, Hadley and Henry, Lionel and PBC, Posit},
+  author    = {Bache, Stefan Milton and Wickham, Hadley and Henry, Lionel and {Posit PBC}},
   year      = {2022},
   month     = mar,
   urldate   = {2024-07-08},
@@ -105,7 +105,7 @@
 
 @misc{chang2024,
   title     = {{Shiny}: {{Web Application Framework}} for {{R}}},
-  author    = {Chang, Winston and Cheng, Joe and Allaire, J. J. and Sievert, Carson and Schloerke, Barret and Xie, Yihui and Allen, Jeff and McPherson, Jonathan and Dipert, Alan and Borges, Barbara and PBC, Posit and Otto, Mark and Thornton, Jacob and Khan, Prem Nawaz and Tsaran, Victor and Lembree, Dennis and Chakravarthula, Srinivasu and O'Connor, Cathy and Petre, Stefan and Rowls, Andrew and Reavis, Brian and Bejaoui, Salmen and Ineshin, Denis and Samhuri, Sami and Fraser, John and Gruber, John and Sagalaev, Ivan and R. Core Team},
[…]
