### 2024-07-08 #6682 @mikemahoney218
No worries! We'd rather get everything right now :smile:

Speaking of which: looking at some R package citations, I'm wondering how you got the list of authors for the packages. For instance, running `citation("shiny")` gives:
```
  @Manual{,
    title = {shiny: Web Application Framework for R},
    author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
    year = {2024},
    note = {R package version 1.8.1.9000, commit 3edf9bfad8ad69a93f744a4f63447442d23e7c9a},
    url = {https://github.com/rstudio/shiny},
  }
```

The paper has a different author list -- I think you've included copyright holders and contributors, when only authors should be included on the citation. This would also be why organizations like Posit, RStudio, R contributors etc are listed as authors -- they generally shouldn't be, and are instead listed on CRAN as non-author roles like copyright holders. This seems to be an issue with most packages; for instance, `rhandsontable` only has one author, but the citation here has about a dozen.

Sorry to not catch this earlier, but would you be able to confirm that packages are only listing authors as authors? Running `citation(<package>)` should give you the proper list.

One other nit: Line 17 -- do you want `Rodwell, 2006` there, or after the word "process"?
