### 2024-07-08 #6682 @ZekeMarshall
Hi @mikemahoney218 ,

Thanks!

Ah, I've grabbed the CRAN package page metadata using the Zenodo web plugin which is giving the complete list of contributors.

I'll work through all my refs and correct the package author names now.

RE the Rodwell, 2006 position, I'll fix that too - thanks!

Cheers,

Zeke
