### 2024-07-08 #6682 @mikemahoney218
Just noticed:
* the Wallace 2018 & 2017 citations need brackets around place names
* The CIEEM citation needs brackets around `april`
* The Marrs citation has a bit of an odd title -- "British National Vegetation Classification Includes software". Should this have a period, or not include the last note?

Sorry to not include these earlier!
