### 2024-07-08 299d1be
More JOSS paper reference fixes
