+++ JOSS/paper.bib
@@ -1,32 +1,27 @@
 @misc{allaire2024,
   title     = {{rmarkdown}: {{Dynamic Documents}} for {{R}}},
-  author    = {Allaire, J. J. and Xie, Yihui and Dervieux, Christophe and McPherson, Jonathan and Luraschi, Javier and Ushey, Kevin and Atkins, Aron and Wickham, Hadley and Cheng, Joe and Chang, Winston and Iannone, Richard and Dunning, Andrew and Yasumoto, Atsushi and Schloerke, Barret and Sievert, Carson and Ryan, Devon and Aust, Frederik and Allen, Jeff and Seo, JooYoung and Barrett, Malcolm and Hyndman, Rob and Lesur, Romain and Storey, Roy and Arslan, Ruben and Oller, Sergio and {Posit PBC} and Otto, Mark and Thornton, Jacob and Farkas, Alexander and Jehl, Scott and Sagalaev, Ivan and Franko, Greg and MacFarlane, John and Raggett, Dave and Gandy, Dave and Sperry, Ben and Drifty and Lister, Aidan and Jonsson, Benct Philip and Krewinkel, Albert},
+  author    = {JJ Allaire and Yihui Xie and Christophe Dervieux and Jonathan McPherson and Javier Luraschi and Kevin Ushey and Aron Atkins and Hadley Wickham and Joe Cheng and Winston Chang and Richard Iannone},
   year      = {2024},
-  month     = may,
-  urldate   = {2024-07-08},
+  note      = {R package version 2.26},
   abstract  = {Convert R Markdown documents into a variety of formats.},
   copyright = {GPL-3},
   doi       = {10.32614/CRAN.package.rmarkdown}
 }
 
 @misc{attali2021,
-  title     = {{shinyjs}: {{Easily Improve}} the {{User Experience}} of {{Your Shiny Apps}} in {{Seconds}}},
-  author    = {Attali, Dean},
-  year      = {2021},
-  month     = dec,
-  urldate   = {2024-07-08},
-  copyright = {MIT + file LICENSE},
-  doi       = {10.32614/CRAN.package.shinyjs}
+  title  = {{shinyjs}: {{Easily Improve}} the {{User Experience}} of {{Your Shiny Apps}} in {{Seconds}}},
+  author = {Dean Attali},
+  year   = {2021},
+  note   = {R package version 2.1.0},
+  doi    = {10.32614/CRAN.package.shinyjs}
 }
 
 @misc{bache2022a,
-  title     = {{magrittr}: {{A Forward-Pipe Operator}} for {{R}}},
-  author    = {Bache, Stefan Milton and Wickham, Hadley and Henry, Lionel and {Posit PBC}},
-  year      = {2022},
-  month     = mar,
-  urldate   = {2024-07-08},
-  copyright = {MIT + file LICENSE},
-  doi       = {10.32614/CRAN.package.magrittr}
+  title  = {{magrittr}: {{A Forward-Pipe Operator}} for {{R}}},
+  author = {Stefan Milton Bache and Hadley Wickham},
+  year   = {2022},
+  note   = {R package version 2.0.3},
+  doi    = {10.32614/CRAN.package.magrittr}
 }
 
 @misc{bsbi2017ellenbergF,
@@ -104,29 +99,25 @@
 }
 
 @misc{chang2024,
-  title     = {{Shiny}: {{Web Application Framework}} for {{R}}},
[…]
