### 2024-07-08 #6682 @ZekeMarshall
Hi @mikemahoney218 ,

I have fixed all of the above (I believe)!

No worries, thanks for taking the time to go over this with such a fine-toothed comb!

Cheers,

Zeke
