### 2024-07-08 #6682 @mikemahoney218
Okay, I think these are my last notes:

* Hill 2007 -- brackets around `british` and `irish` for capitalization
* Smart 2024 -- brackets around `great britain` for capitalization
* Wood 2015 -- brackets around `great britain` for capitalization

After that, I think I'm ready to send this to the EiC for final processing :smile:
