### 2024-07-08 fdee5f8
Fixing caps in hill2007, smart2024, and wood2015 JOSS paper refs
