+++ JOSS/paper.bib
@@ -151,7 +151,7 @@
 @misc{hill2007,
   keywords  = {Resource},
   author    = {Hill, M. O. and Preston, Christopher D. and Bosanquet, Sam D. S. and Roy D. B.},
-  title     = {BRYOATT - Attributes of British and Irish Mosses, Liverworts and Hornworts - Spreadsheet},
+  title     = {BRYOATT - Attributes of {{British}} and {{Irish}} Mosses, Liverworts and Hornworts - Spreadsheet},
   year      = {2007},
   publisher = {{Centre for Ecology and Hydrology}},
   address   = {Huntingdon},
@@ -532,7 +532,7 @@
   keywords  = {biota, environment, biodiversity, plants, trees, woodland, soil, vegetation, ground flora, Environmental survey, Britain},
   language  = {en},
   publisher = {NERC EDS Environmental Information Data Centre},
-  title     = {Broadleaved woodland survey of Great Britain, 2020-2022},
+  title     = {Broadleaved woodland survey of {{Great}} {{Britain}}, 2020-2022},
   url       = {https://catalogue.ceh.ac.uk/id/42c203c8-44de-40e2-a694-b1e8cbd4c8e1},
   year      = {2024}
 }
@@ -601,7 +601,7 @@
 }
 
 @article{wood2015,
-  title     = {Woodland Survey of Great Britain 1971-2001},
+  title     = {Woodland Survey of {{Great}} {{Britain}} 1971-2001},
   volume    = {7},
   issn      = {1866-3516},
   url       = {http://dx.doi.org/10.5194/essd-7-203-2015},
