### 2024-07-08 #6682 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
