### 2024-07-08 #6682 @mikemahoney218
🎉 With everything looking good on my end, it's time for me to hand this back to the EiC for last steps. Thanks @ZekeMarshall  for the submission, and thank you so much to @rasanderson and @adithirgis for reviewing!

Final approvals:
https://github.com/openjournals/joss-reviews/issues/6682#issuecomment-2134574345
https://github.com/openjournals/joss-reviews/issues/6682#issuecomment-2200319428 / https://github.com/openjournals/joss-reviews/issues/6682#issuecomment-2158242408
