### 2024-07-08 #6682 @ZekeMarshall
Hi @mikemahoney218 , just to check... I changed the title of the paper from _RMAVIS v1.0: a R Shiny application..._ to RMAVIS v1.0: a Shiny application... following your advice RE the naming of Shiny, will this be updated by JOSS?
