### 2024-07-08 #6682 @ZekeMarshall
Hi @mikemahoney218 , thanks! However it needs to be changed to **[REVIEW]: RMAVIS v1.0: a Shiny application for the analysis of vegetation survey data and assignment to GB NVC communities** . I'm not sure whether this impacts the paper or is just or housekeeping purposes, just wanted to check!
