### 2024-07-08 #6682 @mikemahoney218
Purely housekeeping 😄 I do need to ask @kthyng (our EiC) if the Zenodo title needs to match the _paper_ title, or just the title of the software, but I think the actual paper title will be pulled from your markdown file.
