### 2024-07-08 #6682 @ZekeMarshall
Good to know! :) The Zenodo title matches the Github repo, but not the JOSS paper title, so that should be fine right? Thanks again for your help and time spent!!
