### 2024-07-08 #6682 @mikemahoney218
Also, I'm realizing now that Kristen is at SciPy this week, so it may be a few days before she's able to take the issue over. Sorry for the delays -- the challenges of getting anything done over the academic summer!
