### 2024-07-09 #6682 @kthyng
@mikemahoney218, @rasanderson I see two unchecked boxes above but it's a bit difficult to wade through the issue to see if it is explained above and anyway there is a strong preference to have all check boxes checked. Do you mind catching me up on the two unchecked boxes?
