### 2024-07-09 #6682 @mikemahoney218
Sorry @kthyng -- I've got links to the final approval comments in https://github.com/openjournals/joss-reviews/issues/6682#issuecomment-2214583654

The final approval from @rasanderson is in https://github.com/openjournals/joss-reviews/issues/6682#issuecomment-2134574345
