### 2024-07-09 #6682 @kthyng
Final approval is a minimum but we'll hope to have the checkboxes checked if possible, to have the perfectionist satisfied in all of us.

In the meantime, to answer your question, the norm is to have the title in the Zenodo metadata match the JOSS paper title (not the software name), though we do not require it.
