### 2024-07-09 #6682 @ZekeMarshall
> Hi! I'll take over now as Track Associate Editor in Chief to do some final submission editing checks. After these checks are complete, I will publish your submission!
>
> * [ ]  Are checklists all checked off?
> * [x]  Check that version was updated and make sure the version from JOSS matches github and Zenodo.
> * [ ]  Check that software archive exists, has been input to JOSS, and title and author list match JOSS paper (or purposefully do not).
> * [ ]  Check paper.

Hi @kthyng , that's great thanks! Please let me know if I need to do anything else.
