### 2024-07-10 #6682 @kthyng
@ZekeMarshall We generally ask authors to update the metadata to match the JOSS paper — do you want to do this or do you prefer not to?
