### 2024-07-10 d0bd219
Updating CITATION.cff file so title matches JOSS paper
