### 2024-07-10 #6682 @ZekeMarshall
HI @kthyng , I've changed the CITATION.cff file so the title matches the paper. I'll also add a preferred citation section containing the JOSS article metadata after publication.

However, just to check. If the idea is that the Zenodo archive is the archive of the software not the JOSS paper and there are two types in the CITATION.cff file: 1) software, and 2) article as I've seen in accepted papers, shouldn't 1) the title in the software citation match the title of the repository and package/app name (e.g. RMAVIS), and 2) the title of the article match the JOSS paper title (e.g. RMAVIS v1.0: a Shiny application for the analysis of vegetation survey data and assignment to GB NVC communities.). Or is this just an artifact of using one CITATION.cff file to serve both Zenodo and JOSS. Apologies if i've missed something!

Please let me know if the changes i've made are ok.
