### 2024-07-17 #6682 @ZekeMarshall
Hi @kthyng ,

Thanks! Though I will defer to the JOSS convention of course!

Please let me know if there is anything else you require.

Zeke
