### 2024-07-26 #6682 @kthyng
Hi @ZekeMarshall and sorry for my terrible delay. There is wiggle room here for sure. The JOSS paper DOI is meant to be the main DOI to cite that comes out of this review process. On the paper is a link to the software archive with Zenodo DOI and we like to have the titles match for consistency. However, sometimes people have a preference for keeping the Zenodo title linked more closely to the software repo itself, and that is fine too. Feel free to set the Zenodo metadata up how you want and let me know when you're finished.
