+++ CITATION.cff
@@ -32,4 +32,4 @@ keywords:
   - NVC
   - Ecology
   - Survey
-title: "RMAVIS v1.0: a Shiny application for the analysis of vegetation survey data and assignment to GB NVC communities."
+title: "RMAVIS"
