### 2024-07-27 #6682 @ZekeMarshall
Hi @kthyng , no worries and thanks for getting back to me! I've changed the CITATION.cff file title to just "RMAVIS". Thanks again!
