### 2024-07-30 1c1ef52
Minor adjustments to JOSS paper
