### 2024-07-30 #6682 @ZekeMarshall
Hi @kthyng , I made two very minor fixes to the JOSS paper, please refer to this most recent pdf!
