### 2024-08-07 #6682 @ZekeMarshall
Hi @kthyng , brilliant thank you very much! Sorry to be a pain, but is there a way to have the paper name displayed as RMAVIS or `RMAVIS` not <code>RMAVIS</code> ?
