### 2024-08-07 #6682 @danielskatz
Hi - Just noticing that the title has encodings in it. That should be changed, after which we can re-accept the paper.  Basically, the backquotes in line 2 of https://github.com/NERC-CEH/RMAVIS/blob/main/JOSS/paper.md should be removed.
