### 2024-08-07 #6682 @ZekeMarshall
Hi @danielskatz , thanks for catching this! I've just removed the title encodings from the paper.md file and pushed. Please let me know if there is anything else I need to do!
