### 2024-08-07 #6682 @danielskatz
> I thought it was breaking things and that is why we were noticing it?

It doesn't break anything exactly, but it creates somewhat bad and ugly metadata.
