### 2024-08-07 d07ace1
Removing period from the end of the JOSS paper title
