+++ JOSS/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'RMAVIS v1.0: a Shiny application for the analysis of vegetation survey data and assignment to GB NVC communities.'
+title: 'RMAVIS v1.0: a Shiny application for the analysis of vegetation survey data and assignment to GB NVC communities'
 tags:
   - Vegetation science
   - NVC
