### 2024-08-07 #6682 @danielskatz
the period wasn't related to the failure - I just happened to notice it when looking at the failure
