### 2024-08-07 #6682 @danielskatz
I *think* the failure is a JOSS one, not related to this. I don't see how trying a re-accept would hurt, though it might fail too
