### 2024-08-07 #6682 @kthyng
Congratulations on your new publication @ZekeMarshall! Many thanks to @mikemahoney218 and to reviewers @adithirgis and @rasanderson for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
